+++ src/pages/createLuckyDay/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useCreateLuckyDayForm } from "./useCreateLuckyDayForm";
